# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KinConnect API
  description: |
    KinConnect is a comprehensive medication management platform that helps patients and caregivers 
    manage medications, track adherence, and coordinate care within family groups.
    
    ## Features
    - **Medication Management**: Create, update, and track medications with reminders
    - **Drug Search**: Search RxNorm database for drug information, interactions, and images
    - **Family Groups**: Invite and manage family members with role-based access
    - **Patient Profiles**: Comprehensive patient health information management
    - **Audit Logging**: Track all access and modifications for compliance
    
    ## Authentication
    All API endpoints require Firebase Authentication using Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```
    
    ## Rate Limiting
    API requests are limited to 100 requests per 15-minute window per IP address.
    
    ## Versioning
    Current API version: v1
    
  version: 1.0.0
  contact:
    name: KinConnect Support
    email: support@kinconnect.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/{project-id}/us-central1/api
    description: Local Development Server
    variables:
      project-id:
        default: kinconnect-dev
        description: Firebase Project ID
  - url: https://us-central1-{project-id}.cloudfunctions.net/api
...
    variables:
...
security:
  - bearerAuth: []
paths:
...
  /patients/profile:
    get:
      tags:
        - Patients
      summary: Get patient profile
      description: Retrieves the authenticated user's patient profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Patient profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Patient'
              examples:
                success:
                  value:
                    success: true
                    data:
                      id: "patient123"
                      userId: "user123"
                      dateOfBirth: "1980-01-01"
                      gender: "male"
                      address: "123 Main St, City, State 12345"
                      phoneNumber: "+1234567890"
                      emergencyContact: "Jane Doe - +1987654321"
                      medicalConditions: ["Hypertension", "Type 2 Diabetes"]
                      allergies: ["Penicillin", "Peanuts"]
                      createdAt: "2024-01-01T00:00:00Z"
                      updatedAt: "2024-01-01T00:00:00Z"
                no_profile:
                  value:
                    success: true
                    data: null
                    message: "No patient profile found"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Patients
      summary: Create patient profile
      description: Creates a new patient profile for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientInput'
            examples:
              basic:
                value:
                  dateOfBirth: "1980-01-01"
                  gender: "male"
                  phoneNumber: "+1234567890"
              complete:
                value:
                  dateOfBirth: "1980-01-01"
                  gender: "male"
                  address: "123 Main St, City, State 12345"
                  phoneNumber: "+1234567890"
                  emergencyContact: "Jane Doe - +1987654321"
                  medicalConditions: ["Hypertension", "Type 2 Diabetes"]
                  allergies: ["Penicillin"]
      responses:
        '201':
          description: Patient profile created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Patient'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Patients
      summary: Update or create patient profile
      description: Updates existing patient profile or creates one if it doesn't exist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientInput'
      responses:
        '200':
          description: Patient profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Patient'
        '201':
          description: Patient profile created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Patient'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== MEDICATIONS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token obtained from Firebase Authentication
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message about the result
        error:
          type: string
          description: Error message if request failed
        data:
          description: Response data (type varies by endpoint)

    # ==================== USER & PATIENT ====================
...
    Patient:
      type: object
      required:
        - id
        - userId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Patient profile ID
        userId:
          type: string
          description: Associated user ID
        dateOfBirth:
          type: string
          format: date
          description: Patient's date of birth
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
          description: Patient's gender
        address:
          type: string
          description: Patient's address
        phoneNumber:
          type: string
          description: Patient's phone number
        emergencyContact:
          type: string
          description: Emergency contact information
        medicalConditions:
          type: array
          items:
            type: string
          description: List of medical conditions
        allergies:
          type: array
          items:
            type: string
          description: List of allergies
        createdAt:
          type: string
          format: date-time
          description: Timestamp when profile was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when profile was last updated
    PatientInput:
      type: object
      properties:
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
        address:
          type: string
        phoneNumber:
          type: string
        emergencyContact:
          type: string
        medicalConditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string

    # ==================== MEDICATION ====================
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            missing_token:
              value:
                success: false
                error: "Authorization header required"
            invalid_token:
              value:
                success: false
                error: "Invalid or expired token"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            server_error:
              value:
                success: false
                error: "Internal server error"